🛢️DB : Change the type of 'answer' from Integer to Long in `Round` entity

+++ backend/src/main/java/com/toyproject/hyeonworld/api/round/domain/dto/in/RoundAnswerCommand.java
@@ -6,7 +6,7 @@ package com.toyproject.hyeonworld.api.round.domain.dto.in;
  */
 public record RoundAnswerCommand (
     long id,
-    int answer
+    long answer
 ){
 
 }
